### 2020-12-16 #2914 @whedon
[PRE REVIEW]: Visualization of Multi-Dimensional Data - The data-slicer Package
**Submitting author:** @kuadrat (<a href="http://orcid.org/0000-0001-5523-6924">Kevin Kramer</a>)
**Repository:** <a href="https://github.com/kuadrat/data_slicer" target ="_blank">https://github.com/kuadrat/data_slicer</a>
**Version:** 0.0.9
**Reviewers:** @sabinomaggi  @Chilipp
