### 2020-12-17 #2914 @kuadrat
>
> @kuadrat if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
>

Just judging from that list (starting at the bottom), the following people might be well suited as reviewers:
sabinomaggi
jodemaey
kgoebber
Generally people with a background in data heavy natural sciences should be suitable.
A background in condensed matter physcics (x-ray/neutron/synchrotron science) may help understand where we are coming from, but should not be necessary.

Kind regards
