### 2020-12-17 #2914 @jgostick
This needs a 'scope query'.  It's very light, despite having 2k lines of python.  It a jazzed up jupyter notebook for looking at 3d data from different directions.  If it does more than this, it's not documented on the website.
