### 2020-12-18 #2914 @kthyng
Ok let's do this. We'll scope query based on @jgostick's comment, and I will also add @timtroendle as the handling editor for when back from being out of office. If the submission passes the scope query, we'll have our editor in place.

@kuadrat It's possible you'll hear back about the scope query next week, but it may be after the holidays at this point. Thanks for your patience.
