### 2020-12-18 #2914 @kthyng
@openjournals/dev sorry for the trouble, but can I remove the editor I added?
