### 2021-01-04 #2914 @timtroendle
👋 @sabinomaggi, @jodemaey, @kgoebber would any of you be willing to review this submission for JOSS?
