### 2021-01-04 #2914 @sabinomaggi
Seems very interesting. I can review the paper but not earlier than 3-4 weeks. Am I supposed to test the package to check how it works?
