### 2021-01-05 #2914 @timtroendle
Thanks @sabinomaggi , that's great to hear. We are currently in [Reduced service mode](https://blog.joss.theoj.org/2020/05/reopening-joss), which means that reviews may take up to six weeks. Of course, we are encouraging reviewers to start the review as soon as possible.

Regarding your question: the review includes a functionality check as well. You may want to have a quick look at our [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html).
