### 2021-01-11 #2914 @timtroendle
👋 @Chilipp, @krischer, would any of you be willing to review this submission for JOSS?
