### 2021-01-19 #2969 @whedon
[REVIEW]: Visualization of Multi-Dimensional Data - The data-slicer Package
**Submitting author:** @kuadrat (<a href="http://orcid.org/0000-0001-5523-6924">Kevin Kramer</a>)
**Repository:** <a href="https://github.com/kuadrat/data_slicer" target ="_blank">https://github.com/kuadrat/data_slicer</a>
**Version:** 0.0.9

@sabinomaggi,  @Chilipp, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
