### 2021-01-25 #2969 @sabinomaggi
Review accepted. It seems I cannot compile te paper due to a missing date in the paper yaml header. May I edit the
data_slicer/paper/paper.md file myself or only the author is supposed to do these changes?
