### 2021-01-25 #2969 @kuadrat
Hi, and thanks in advance for taking the time to review my submission.
I just took the liberty of putting today's date in the header (I assume the date should be adjusted at a later time?).
Does it compile for you now?
Sorry about the inconvenience. The paper did compile when I tested it with [the github action](https://github.com/kuadrat/data_slicer/actions/runs/509304889) before submission.
