### 2021-01-25 #2969 @sabinomaggi
@kuadrat Hi, yes as you can see the paper compiles fine now. In the meantime I was able to get a decent version of the paper with all references by using pandoc and xelatex, but this automatic compilation is way better than a manual approach.
I am really eager to read the paper and test the package in full.
