### 2021-02-08 kuadrat/data_slicer#3 @kuadrat
Can't get attribute `ds_cmap`
Issue originally formulated by @sabinomaggi

Hi Kevin, I have followed the installation instructions in https://data-slicer.readthedocs.io/en/latest/installation.html and everything seems OK.

    (base) $ conda create --name dataslicer python==3.7.5
    (base) $ conda activate dataslicer
    (dataslicer) $ pip install --upgrade pip
    [...]
    Successfully installed pip-21.0.1

    (dataslicer) $ pip install data_slicer
    Collecting data_slicer
    [...]
    Installing collected packages: ipython-genutils, wcwidth, traitlets, six, [...]
    Successfully installed [...] data-slicer-0.0.9 [...]

However, when I try to test the installation by running pit I get this error message

```
(dataslicer) $ pit
===== 2021.02.08 14:02:05 =====
Traceback (most recent call last):
  File "/Users/maggi/anaconda3/envs/dataslicer/bin/pit", line 5, in <module>
    from data_slicer.pit import start_main_window
  File "/Users/maggi/anaconda3/envs/dataslicer/lib/python3.7/site-packages/data_slicer/pit.py", line 75, in <module>
    cmaps = pickle.load(f)
AttributeError: Can't get attribute 'ds_cmap' on <module '__main__' from '/Users/maggi/anaconda3/envs/dataslicer/bin/pit'>
```

I have also repeated this precedure using the latest available version of Python and the error is exactly the same.

I have not been able to find any info about package compatibility in the documentation (and also in the paper under review), and therefore it is possible that data-dlicer runs only under Linux. However, if this is true, I think you should explicitly state this fact in the documentation, as well as in the paper (but this last point is clearly part of the review).
If, on the other hand, data-dlicer is supposed to run also under macOS, could you please help to solve this issue?
