### 2021-02-08 #2969 @kuadrat
Hi @sabinomaggi,

thanks for the detailed, calm and polite feedback!
If I understand it correctly, problems like these during JOSS review should be opened as Github issues directly in the repo under review (I may be wrong, @timtroendle may want to correct me).
I have taken the liberty of opening the problem you mention as an [issue ](https://github.com/kuadrat/data_slicer/issues/3) and will reply more specifically there.

Best regards,
-Kevin
