### 2021-02-08 kuadrat/data_slicer#3 @kuadrat
Hi @sabinomaggi and thanks again for the detailed feedback!

Indeed, compatibility information is missing and should be there. I will add it to the readme and documentation.
In principle, the software should run on Linux, Windows and macOS.

I actually suspect that the error here arises due to a last minute feature that I have implemented shortly before the submission to JOSS. Namely a form of colormap caching, intended to decrease startup time.

To be more precise, in an older version, all the colormap objects would be created on startup. I noticed that this took up time in the order of seconds, which is not ideal, especially since usually not many colormaps will be used during a session. So I thought instead of recreating them on each startup, I could just load previously `pickle`d objects - which eventually seemed to work but only after a lot of trail and error. There's likely something unstable with the way it's implemented now, still.

I'll see if I can fix it, but suspect that I will likely just have to remove the feature and live with longer start up times for now.
