### 2021-02-08 kuadrat/data_slicer#3 @sabinomaggi
Hi @kuadrat , thanks for the quick reply. If you want I can clone the repo and go back to a previous commit to check if your assumption is true (at least on macOS).
