### 2021-02-08 kuadrat/data_slicer#3 @kuadrat
So my first hunch is probably correct. This is not a compatibility issue, but a proper bug.
I didn't notice it, because when testing, I would start pit using `python -m data_slicer.pit` instead of the more convenient `pit` entry point.
Maybe I can find out why the unpickling works in one case but not in the other.

Can you check if `python -m data_slicer.pit` works for you?
