### 2021-02-08 kuadrat/data_slicer#3 @kuadrat
> Hi @kuadrat , thanks for the quick reply. If you want I can clone the repo and go back to a previous commit to check if your assumption is true (at least on macOS).

That would be a pathway we could follow if `python -m data_slicer.pit` doesn't work.
The commit in question would be [be5d0991f3b20080b0b66548a36b6b91d536a79a](https://github.com/kuadrat/data_slicer/commit/be5d0991f3b20080b0b66548a36b6b91d536a79a). Most commits after that pertain to clean up and paper writing, so we shouldn't miss any essential features by working with that version.
