### 2021-02-08 d052d20
compatibility info in doc. Logo.
