### 2021-02-08 f84779c
Revamped cmap caching. Now "on demand".
