### 2021-02-08 kuadrat/data_slicer#3 @kuadrat
Thanks to your input I was able to think of a much more straightforward way to handle caching that should not lead to the issue described.
If you could test again with the newest version (0.0.10), I'd be happy to hear the results.
