### 2021-02-08 #2969 @sabinomaggi
@timtroendle Ok. This is the [link to the issue](https://github.com/kuadrat/data_slicer/issues/3), and as far as I can say it will be solved very soon.
