### 2021-02-08 kuadrat/data_slicer#3 @sabinomaggi
@kuadrat
`python -m data_slicer.pit`
works and all the elements of the interface respond to my mouse clicks (now I must discover how to use my own data).
