### 2021-02-08 kuadrat/data_slicer#3 @sabinomaggi
> Thanks to your input I was able to think of a much more straightforward way to handle caching that should not lead to the issue described.
> If you could test again with the newest version (0.0.10), I'd be happy to hear the results.

Of course yes, takes nothing...
```
$ pip install data_slicer --upgrade
[...]
Installing collected packages: data-slicer
  Attempting uninstall: data-slicer
    Found existing installation: data-slicer 0.0.9
    Uninstalling data-slicer-0.0.9:
      Successfully uninstalled data-slicer-0.0.9
Successfully installed data-slicer-0.0.10
```
You are right, now `pit` works perfectly.
