### 2021-02-15 kuadrat/data_slicer#3 @sabinomaggi
I think this issue can be closed, now.
