### 2021-02-18 #2969 @timtroendle
@sabinomaggi, that's great to hear. I seen you have started ticking the boxes in the checklist above and I assume you have comments to all unticked boxes. Could you open issues in the package repository for the things that need changes in your opinion? And could you mention this issue here in those issues, so as to link the issues?
