### 2021-02-18 #2969 @timtroendle
@sabinomaggi  Exactly, but you do not need to open an individual issue per item on the checklist. It's up to you how fine-grained you believe the issues need to be.
