### 2021-02-18 kuadrat/data-slicer#4 @sabinomaggi
Data formats compatible with data_slicer
Hi @kuadrat one of the main shortcoming of the documentation is the absence of any description about the format of the data that `data-slicer` (and/or `PIT`) can process, as well as about how make PIT read a generic datafile besides those distributed with the package.
A discussion on how to convert one's own data to a format that data-slicer/PIT can read could also be a welcome addition.
And since we are talking about files, why not to allow PIT to read a data file from the command line?

**Disclaimer**

This issue is part of the JOSS review in [https://github.com/openjournals/joss-reviews/issues/2969](https://github.com/openjournals/joss-reviews/issues/2969)
