### 2021-02-18 #2969 @sabinomaggi
@timtroendle @kuadrat

In this paper the authors present a new Python package, `data-slicer`, whose main goal is to provide fast visualization of 3D data. Unfortunately, the actual presentation of the `data-slicer` package is too short to be useful (about 6 lines of text) and the authors quickly focus on the graphical tool built upon the package (`PIT`), which looks as the real focal point of the paper. If this is true, the title and the content of the paper should clearly reflect this fact. Otherwise, the authors should definitely enlarge and improve the presentation of the `data-slicer` package.

In a very short article like those submitted to JOSS this lack of a clear focus could be a minor problem, provided that the repository of the package gave all the necessary information. However, the same terseness found in the paper characterizes the GitHub repository of `data-slicer`, which lacks to illustrate the different functions of the `data-slicer` package and how they can be put together to build scripts to visualize one's own data.

Instead, the available documentation focuses – in my opinion too heavily – on `PIT`, and again gives the impression that the main goal of this work is the presentation of the GUI application built on top if `data-slicer`, and not of the package itself.

Another clear shortcoming of the documentation ([link to the issue](https://github.com/kuadrat/data_slicer/issues/4)) is the absence of any description about the format of the data that `data-slicer` (and/or `PIT`) can process, as well as about how make `PIT` read a generic datafile besides those distributed with the package. As the main goal of `data-slicer`/`PIT` is to provide a tool for _"3D data visualization and manipulation"_, the discussion about the supported file formats and how one can convert his/her own data to a format that `data-slicer`/`PIT` can read should be one of the central points of the documentation. It took me a couple of hours just to hack a script that created a [simple 3D dataset](https://user-images.githubusercontent.com/582340/108406846-3a272d80-7223-11eb-9235-dfb74a20a683.png) that `PIT` could read, and to do that I needed to refer to the source code of the package. On the positive side, after having done that I was able to make `PIT` read and process fairly large datafiles (up to 0.5 GB) without any hiccups.

The documentation also fails to mention if `PIT` can read a generic datafile when launched from the command line ([link to the issue](https://github.com/kuadrat/data_slicer/issues/4)), for example, by writing the filename after the `pit` command (I have tried and it seems it is not possible). As a reviewer I thought it was worthwile to invest some hours to better understand how `PIT` worked, but I don't know if normal users can share the same attitude. I understand that a package such as `data-slicer` is not for the casual user, but better and detailed documentation can be only beneficial to spread the usage of a package like this.

That being said, I truly believe that `data-slicer` has a great potential and could become a very interesting and useful Python package (and this feeling was the main reason to review this JOSS paper) once the authors/developers better describe the package and how to use it, both in the repository and in the paper.

Minor remarks:

- The package is named `data-slicer` in the paper (see below for an exception) but `data_slicer` in the GitHub repository. Even if this is just a minor cosmetic problem, fixing this ambiguity from the beginning would avoid troubles later (e.g., with Google indexing algorithms).
- I am not a specialist of visualization of 3D data and I could be wrong, but I wonder why the authors did not mention, neither in the paper nor in the references, well-known and widely available visualizers such as [ParaView](https://www.paraview.org), [Mayavi](http://code.enthought.com/pages/mayavi-project.html) or [VisIt](https://wci.llnl.gov/simulation/computer-codes/visit) (and possibly also the historic but now  seemingly defunct [OpenDX](https://en.wikipedia.org/wiki/IBM_OpenDX) package).
- The reference to Figure 1 (line #12) should briefly explain the actual content of the figure (e.g, that the Figure compares the measured data plots published in the last decades of the XX century with the corresponding complex datasets produced by modern instrumentation). I also think that the references should preferably go here and not in the caption of Figure 1..
- The caption of Figure 1 should be rewritten according to the changes in the text.
- Figure 2: There is no reference to this Figure in the text of the paper.
- Figure 2: The title "data_slicer package" is not needed (and in all cases, according to the remainder of the text, it should read "data-slicer package"). Also, if `PIT` is only an example of a tool build using the package functions, I think it should go outside the main box containing the `data-slicer` building blocks.
- The sentence "The package can be installed from pypi using `pip install data_slicer`" in the README file of the package should read "The package can be installed from PyPI using `pip install data_slicer`", preferably with a link to the Python Package Index home page (I am not sure that all scientists potentially interested to `data[-_]slicer` know the existence of the PyPI repository).
- English should be improved (see for example lines no. 20, 41-42, 43-44, 47 or 51).
