### 2021-02-22 kuadrat/data_slicer#4 @kuadrat
Hi @sabinomaggi and thanks for your feedback. Judging from the more detailed description in your review report, I could fully understand if this lack of documentation on my side has let to a considerable amount of frustration - I know it would have for me. So I want to thank you for your dedication in finding the way to do it from the source code and apologize for the hassle.

Since all testers so far have been using compatible data formats to begin with and documentation principally exists ([though it is sparse and not at all obvious where](https://data-slicer.readthedocs.io/en/latest/data_slicer.html#data_slicer.dataloading.Dataloader_3dtxt)), this issue has flown under the radar.

I'm going to add a section to a more visible part of the documentation.
