### 2021-02-22 kuadrat/data_slicer#4 @sabinomaggi
@kuadrat This is exactly what I tried to do at the beginning (I usually prefer text data files over binary formats), but it was very cumbersome to do and at the end I decided that for my own testing a pickle was far easier and faster.
The problem is that searching through an API is not a job that everyone can do. I think it would be far better for the success of your package to write one (or more) tutorials describing how to handle data sets and how to put together the many data-slicer functions to visualize one's own data. My two cents...
