### 2021-02-22 de8e38a
Fix cmap caching related bug
