### 2021-02-22 4045c38
pickle dataloading convenience methods
