### 2021-02-22 2f8de5b
fixed tests (bug resulting from changed cmap caching)
