### 2021-02-22 5d6b22c
Added MRI brain scan example data
