+++ doc/source/data_format.rst
@@ -1,0 +1,217 @@
+.. _sec-data_format:
+
+Data formats and data loading
+=============================
+
+The existing dataloaders found in the :mod:`dataloading module 
+<data_slicer.dataloading>` can handle two types of input data formats:
+
+1. A binary file that has been created using python's ``pickle`` module.
+   The pickled object can be either a ``numpy`` array, or a dictionary or 
+   :class:`argparse.Namespace` containing the keys *data* and *axes*, as 
+   described in :class:`~data_slicer.dataloading.Dataloader_Pickle`.
+
+2. A plain text file containing values for x, y, z and the actual data in 
+   four columns as described in the documentation of 
+   :class:`~data_slicer.dataloading.Dataloader_3dtxt`.
+   Notice that there is an utility function 
+   (:func:`~data_slicer.dataloading.three_d_to_txt`) that can help you 
+   create such a text file in the correct format from existing data.
+
+In the following we give an example tutorial for both cases.
+
+
+Binary pickle files
+-------------------
+
+Step 1: Load the data into python
+.................................
+
+We will assume our starting point to be a 3D data file of any format.
+It could be an `Igor file <https://www.wavemetrics.com/igor-8-highlights>`_, 
+an `HDF5 file <https://www.hdfgroup.org/solutions/hdf5>`_, a `matlab 
+file <https://www.mathworks.com/products/matlab.html>`_, a plain text file but 
+in a different format than what is required by 
+:class:`~data_slicer.dataloading.Dataloader_3dtxt` or anything else.
+In any case, you will have to find a way of loading that dataset into 
+python first
+(For the examples listed above, the packages `igor 
+<https://pypi.org/project/igor/>`_, `h5py <https://github.com/h5py/h5py>`_, 
+`scipy 
+<https://docs.scipy.org/doc/scipy/reference/generated/scipy.io.loadmat.html>`_ 
+or `numpy 
+<https://numpy.org/doc/stable/reference/generated/numpy.loadtxt.html>`_ could 
+be used, respectively).
+
+In order to make the example more concrete and to allow for a step-by-step 
+code-along experience, we will explain here how we prepared the MRI brain scan 
+data that can be loaded by issuing ``mw.brain()`` on ``PIT``'s console.
+But again, how exactly this step is handled depends on your starting point.
+The result of Step 0 should always be the same, though: Your data should be 
+accessible as a numpy array in python.
+
+OK, now let's get started.
+We find and download our brain scan data of a meditating person 
+`here <https://openneuro.org/crn/datasets/ds000108/snapshots/00002/files/sub-01:anat:sub-01_T1w.nii.gz)>`_.
+
+Once downloaded, we need a way of opening it.
+A quick `Ecoisa search <www.ecosia.org>`_ leads us to the library `NiBabel 
+<https://nipy.org/nibabel/gettingstarted.html>`_, which seems to be able to 
+open ``nii`` files.
+Thus, we install that library (depending on your system and setup there may 
+be different ways of doing this)::
+
+   pip install nibabel
+
+Now, following the instructions on the NiBabel webpage, we load the image 
+data as a numpy array::
+
+   python
+   >>> import nibabel as nib
+   >>> img = nib.load('sub-01_T1w.nii.gz')
+   >>> my_data = img.get_fdata()
+
+.. note::
+   You need to be in the directory where you placed the downloaded file (here 
+   ``sub-01_T1w.nii.gz``) in order for this to work.
+
+.. note::
+   Just to point it out once more, the details of this first step depend very 
+   much on your use case. It also does not matter whether you work in the 
+   live python interpreter like in the example or whether you wrap it all in 
+   a script.
+   You're fine as long as you have a way of getting your data into the form 
+   of a numpy array.
+
+Step 2: Optionally change data arrangement
+..........................................
+
[…]
