+++ doc/source/plugins.rst
@@ -1,3 +1,5 @@
+.. _sec-plugin:
+
 Plugins
 =======
 
