### 2021-02-22 9aca02f
Allow giving a file as argument to calling pit from command line
