### 2021-02-23 41b0df8
doc: typos and small fixes
