+++ setup.py
@@ -8,7 +8,8 @@ with open("README.md", "r") as fh :
 
 setuptools.setup(
     name='data-slicer',
-    version='0.0.9',
+#    version='99.0.2',
+    version='0.0.12',
     author='Kevin Kramer',
     author_email='kevin.kramer@uzh.ch',
     description='Tools for quick visualization of three dimensional datasets.',
