### 2021-02-23 kuadrat/data-slicer#4 @kuadrat
Hi @sabinomaggi,
> I think it would be far better for the success of your package to write one (or more) tutorials  describing how to handle data sets and how to put together the many data-slicer functions to visualize one's own data.

I absolutely agree.
Explanations and a tutorial on data formats have been added [here](https://data-slicer.readthedocs.io/en/latest/data_format.html).
Also, there is now some more intuitive example data that can be loaded with ``mw.brain()``, referenced in two places in the documentation.
I plan on extending the documentation with an additional step-by-step guide on working with this example data.

Furthermore, the newest version (0.0.12) gives you the functionality that you rightfully expected, namely loading a file directly when launching ``PIT`` from the command line.
