+++ paper/paper.bib
@@ -1,3 +1,23 @@
+@online{00mayavi,
+  url = {https://docs.enthought.com/mayavi/mayavi/},
+  urldate = {2021-02-25},
+  title = {Mayavi: {{3D}} Scientific Data Visualization and Plotting in {{Python}} \textemdash{} Mayavi 4.7.2 Documentation}
+}
+
+@online{00visit,
+  url = {https://wci.llnl.gov/simulation/computer-codes/visit},
+  urldate = {2021-02-25},
+  title = {{{VisIt}} | {{Weapons}} and {{Complex Integration}}}
+}
+
+@article{ahrens05paraview,
+  shorttitle = {{{ParaView}}},
+  author = {Ahrens, J. and Geveci, Berk and Law, Charles},
+  date = {2005-01-01},
+  journaltitle = {Visualization Handbook},
+  title = {{{ParaView}}: {{An End}}-{{User Tool}} for {{Large Data Visualization}}}
+}
+
 @online{3dslicer,
   url = {https://www.slicer.org/},
   urldate = {2020-12-01},
