### 2021-02-25 kuadrat/data-slicer#5 @kuadrat
Documentation on combining widgets & functions
As correctly pointed out by @sabinomaggi, this is clearly missing and users have no chance of figuring out how to do this without diving into the API reference or the source code itself.
