### 2021-02-25 #2969 @kuadrat
@sabinomaggi

Thanks again for your thorough review and these constructive comments.
It is clear to me that addressing them will improve the paper and the software package significantly.

In the following, I will try to respond to your concerns point by point.

> In this paper the authors present a new Python package, `data-slicer`, whose main goal is to provide fast visualization of 3D data. Unfortunately, the actual presentation of the `data-slicer` package is too short to be useful (about 6 lines of text) and the authors quickly focus on the graphical tool built upon the package (`PIT`), which looks as the real focal point of the paper. If this is true, the title and the content of the paper should clearly reflect this fact. Otherwise, the authors should definitely enlarge and improve the presentation of the `data-slicer` package.
>
> In a very short article like those submitted to JOSS this lack of a clear focus could be a minor problem, provided that the repository of the package gave all the necessary information. However, the same terseness found in the paper characterizes the GitHub repository of `data-slicer`, which lacks to illustrate the different functions of the `data-slicer` package and how they can be put together to build scripts to visualize one's own data.
>
> Instead, the available documentation focuses – in my opinion too heavily – on `PIT`, and again gives the impression that the main goal of this work is the presentation of the GUI application built on top if `data-slicer`, and not of the package itself.

The observation that both, the paper and the package documentation are focused on `PIT` is correct.
The reason for this is that `PIT` is actually what has been demanded and used by the current userbase and is expected to be the starting point for most future users.
However, a strength of the modular design lies in the possibility for a programmatically inclined user to create new applications from the provided widgets.
Since this possibility should not be missing, the scope of the paper was chosen to present the full `data-slicer` as a package, with a focus on `PIT`.
At this point I still feel that this framing makes sense but fully agree that description and documentation on how to make use of these widgets is lacking.
My thinking was that the users who would potentially make use of this feature would be consulting the [code reference](https://data-slicer.readthedocs.io/en/latest/data_slicer.html), but it is clear that more explicit explanations and tutorials are needed.
I would therefore propose that I keep the framing as it is, but to vastly extend explanations of this modular approach in the documentation.
I have created [an issue](https://github.com/kuadrat/data-slicer/issues/5#issue-816304658) for tracking progress on this front.

> Another clear shortcoming of the documentation ([link to the issue](https://github.com/kuadrat/data_slicer/issues/4)) is the absence of any description about the format of the data that `data-slicer` (and/or `PIT`) can process, as well as about how make `PIT` read a generic datafile besides those distributed with the package. As the main goal of `data-slicer`/`PIT` is to provide a tool for _"3D data visualization and manipulation"_, the discussion about the supported file formats and how one can convert his/her own data to a format that `data-slicer`/`PIT` can read should be one of the central points of the documentation. It took me a couple of hours just to hack a script that created a [simple 3D dataset](https://user-images.githubusercontent.com/582340/108406846-3a272d80-7223-11eb-9235-dfb74a20a683.png) that `PIT` could read, and to do that I needed to refer to the source code of the package. On the positive side, after having done that I was able to make `PIT` read and process fairly large datafiles (up to 0.5 GB) without any hiccups.
>
> The documentation also fails to mention if `PIT` can read a generic datafile when launched from the command line ([link to the issue](https://github.com/kuadrat/data_slicer/issues/4)), for example, by writing the filename after the `pit` command (I have tried and it seems it is not possible). As a reviewer I thought it was worthwile to invest some hours to better understand how `PIT` worked, but I don't know if normal users can share the same attitude. I understand that a package such as `data-slicer` is not for the casual user, but better and detailed documentation can be only beneficial to spread the usage of a package like this.

Both of these valid concerns are being addressed in the linked issue.

> That being said, I truly believe that `data-slicer` has a great potential and could become a very interesting and useful Python package (and this feeling was the main reason to review this JOSS paper) once the authors/developers better describe the package and how to use it, both in the repository and in the paper.
>

Thank you again for the encouragement and constructive feedback!
It has helped me revisit the paper and realize that I had been struck with severe tunnel vision after having worked with the software so closely for so long that I have indeed not done a particularly good job at explaining it to outsiders.
I am confident that the revised version will present a significant improvement.

## Addressing minor remarks

> Minor remarks:
>
>     * The package is named `data-slicer` in the paper (see below for an exception) but `data_slicer` in the GitHub repository. Even if this is just a minor cosmetic problem, fixing this ambiguity from the beginning would avoid troubles later (e.g., with Google indexing algorithms).
>

Thanks for the tip. The repository has been renamed to be consistent with the name on PyPI.

>     * I am not a specialist of visualization of 3D data and I could be wrong, but I wonder why the authors did not mention, neither in the paper nor in the references, well-known and widely available visualizers such as [ParaView](https://www.paraview.org), [Mayavi](http://code.enthought.com/pages/mayavi-project.html) or [VisIt](https://wci.llnl.gov/simulation/computer-codes/visit) (and possibly also the historic but now  seemingly defunct [OpenDX](https://en.wikipedia.org/wiki/IBM_OpenDX) package).
>

These references definitely belong in the paper, and have been added.
We simply did not have them on the radar as they are rarely (if at all) used in our community.
The latter two, in particular, seem to be more suited to creating beautiful 3D renderings of data.
For users who want to have a way of quickly inspecting their datasets - potentially during a running experiment - these powerful tools become unhandy to wield.
I have reformulated large parts of the `Statement of Need` section to make this more clear.

>     * The reference to Figure 1 (line no. 12) should briefly explain the actual content of the figure (e.g, that the Figure compares the measured data plots published in the last decades of the XX century with the corresponding complex datasets produced by modern instrumentation). I also think that the references should preferably go here and not in the caption of Figure 1.
>
>     * The caption of Figure 1 should be rewritten according to the changes in the text.
>

The sentence in the main text leading to Figure 1 has been made more explicit.

Personally, I believe it is better if the references and explanations are as close to the referenced material as possible.
But maybe this is a question of taste and should be left to editorial guides or the editor @timtroendle to decide.

>     * Figure 2: There is no reference to this Figure in the text of the paper.
>

A sentence has been added in the section `Summary`.

>     * Figure 2: The title "data_slicer package" is not needed (and in all cases, according to the remainder of the text, it should read "data-slicer package"). Also, if `PIT` is only an example of a tool build using the package functions, I think it should go outside the main box containing the `data-slicer` building blocks.
>

The large grey block is meant to represent the `data-slicer` package as a whole, so labeling it as such seemed sensible (the label has been made consistent).
Since `PIT` ships with the `data-slicer` package, I think it makes sense to have it inside the box.

However, realizing that I may be suffering from tunnel view, I am happy to enact your proposed changes if you still find it difficult to understand the figure.
After all, if it creates more confusion than it resolves, it certainly does not serve its purpose.

>     * The sentence "The package can be installed from pypi using `pip install data_slicer`" in the README file of the package should read "The package can be installed from PyPI using `pip install data_slicer`", preferably with a link to the Python Package Index home page (I am not sure that all scientists potentially interested to `data[-_]slicer` know the existence of the PyPI repository).
>

That's a useful observation and your suggestion has been implemented.

>     * English should be improved (see for example lines no. 20, 41-42, 43-44, 47 or 51).

Sentences have been simplified and uncommon formulations removed.
