+++ paper/paper.bib
@@ -1,13 +1,13 @@
 @online{00mayavi,
   url = {https://docs.enthought.com/mayavi/mayavi/},
   urldate = {2021-02-25},
-  title = {Mayavi: {{3D}} Scientific Data Visualization and Plotting in {{Python}} \textemdash{} Mayavi 4.7.2 Documentation}
+  title = {Mayavi}
 }
 
 @online{00visit,
   url = {https://wci.llnl.gov/simulation/computer-codes/visit},
   urldate = {2021-02-25},
-  title = {{{VisIt}} | {{Weapons}} and {{Complex Integration}}}
+  title = {VisIt}
 }
 
 @article{ahrens05paraview,
