### 2021-02-25 kuadrat/data-slicer#4 @Chilipp
hey @kuadrat! thank you very much for adding this content. I agree with @sabinomaggi that this is important.

> Explanations and a tutorial on data formats have been added here.
> Also, there is now some more intuitive example data that can be loaded with mw.brain(), referenced in two places in the documentation.

I also recommend to include how to plot data directly, without using a local pickle or text file. From what I got from your code, this can be done via something like `pit.prepare_data(my_data, axes)`?

I found another issue with your new documentation that I'll reference in another issue, though
