### 2021-02-25 kuadrat/data-slicer#6 @Chilipp
Wrong axis specification in docs Quick start
hey @kuadrat!

this issue is a quick follow-up on #4.

In your docs you state, that axes should be encoded such as

```python
    D = dict(data=my_data, axes=my_axes)
    dump(D, 'brain.p')
```

where `my_axes` is defined as

```python
my_axes = [x_axis, y_axis, z_axis]
```

however, this produces the following error for the pickle dataloader:
`Exception <class 'KeyError'>: 'xaxis'`. It works, if you define `D` instead
as

```python
D = dict(data=my_data, xaxis=x_axis, yaxis=y_axis, zaxis=z_axis)
```

### Disclaimer
this issue is part of the JOSS review in https://github.com/openjournals/joss-reviews/issues/2969
