### 2021-02-25 kuadrat/data-slicer#7 @Chilipp
add a statement of need to the documentation
to fulfill the requirements of JOSS, you need to include

> **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I could not find this in the docs, only in the `paper.md` file

### Disclaimer
this issue is part of the JOSS review in https://github.com/openjournals/joss-reviews/issues/2969
