### 2021-02-25 kuadrat/data-slicer#8 @Chilipp
document your dependencies
according to JOSS, you must clearly sate your list of dependencies.

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

You already include them in your `requirements.txt` and `setup.py` file

```
matplotlib>=3.1.1
numpy>=1.17.3
PyOpenGL>=3.1.1a1
pyqtgraph>=0.11.0
PyQt5>=5.13.2
qtconsole>=4.6.0
```

but you should also document them in the docs to (1) give your package users the chance to see the necessary requirements before installing the package, and (2) acknowledge the open-source frameworks that your software is built upon

### Disclaimer
this issue is part of the review for JOSS in https://github.com/openjournals/joss-reviews/issues/2969
