### 2021-02-25 kuadrat/data-slicer#9 @Chilipp
add tests
I could not find any automated tests or documented steps to verify the functionality of the software, but this is a requirement for JOSS

> **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

### Disclaimer
this issue is part of the review for JOSS in https://github.com/openjournals/joss-reviews/issues/2969
