### 2021-02-25 kuadrat/data-slicer#10 @Chilipp
add community guidelines and a developer guide
according to the JOSS review criteria, you need

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I could not find them in the docs or the source code repository. see https://docs.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors for some help

### Disclaimer
this issue is part of the review in https://github.com/openjournals/joss-reviews/issues/2969
