### 2021-02-25 kuadrat/data-slicer#5 @Chilipp
hey @kuadrat! I totally agree with you and @sabinomaggi that this is necessary, especially as you claim this flexibility of your framework in the JOSS paper.
