### 2021-02-25 kuadrat/data-slicer#11 @Chilipp
add missing references
as part of the review in https://github.com/openjournals/joss-reviews/issues/2969, I am missing some important references in your paper.

You should cite the software packages that your using (if possible), namely matplotlib and
numpy. Please lookup the corresponding citation instructions in their official
documentation.
