### 2021-02-25 #2969 @Chilipp
hey @kuadrat and @timtroendle, my honest apologies for the delay in my review. Too many unexpected things happend in the last 4 weeks. Anyway, I went through the checklist and created a bunch of issues for the individual items that I think are missing. @kuadrat, please refer to these issues when you are making commits or PRs to your source code repository to fix them (this would make my life a lot easier :sweat_smile:).

All in all I have to say that your UI design is quite intuitive (in combination with the quick start document) and I can see multiple potential use cases for it. It has a lot of potentials also for other disciplines. The main issue that I see is the lack of automated tests or steps to verify the functionality of the software (see https://github.com/kuadrat/data-slicer/issues/9).

@timtroendle, I have one issue where I'd like to get your advice

> **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the JOSS guidelines

the software is quite new and the functionality and direct link to research is rather limited due to it's generality. But I saw that you already made a [*scope query*](https://github.com/openjournals/joss-reviews/issues/2914#issuecomment-747727385) whether this submission fulfills the criteria so I assume I can checkoff this item?
