### 2021-02-25 #2969 @Chilipp
As a side note: I am working more in the field of earth and climate sciences, but in fact, slicing data cubes is a common use case in my field, too. So there is quite some potential if we can add the possibility for georeferenced data (e.g. via [`cartopy`](https://scitools.org.uk/cartopy/docs/latest/) and a dataloader that supports netCDF files (e.g. via [`xarray`](http://xarray.pydata.org/en/stable/)). But this has of course nothing to do with this review, just as a personal remark :smiley:
