### 2021-02-26 #2969 @timtroendle
@Chilipp Thank you for your review! Regarding your question: yes, there was an editorial scope check before. Still, we'd be interested in your opinion. Scholarly effort does not necessarily require to have been used or cited before, [but there are other criteria as well](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort).
