### 2021-02-26 #2969 @sabinomaggi
@Chilipp I totally agree with your remark. I think it would be very interesting to use this package to have a quick look to satellite imagery data, e.g. to visualize multispectral or hyperspectral data corresponding to each pixel. This has nothing to do with this review, but could be a possible direction for future work.
