### 2021-02-26 kuadrat/data-slicer#12 @sabinomaggi
Add information for updating the package
Hi, while updating the package to 0.0.12 I realized that it would be useful to add a couple of lines in the documentation (https://data-slicer.readthedocs.io/en/latest/installation.html) to instruct users on how to update the package.
Nothing that cannot be easily found with a quick Google search, but easing the life of users is a nice touch.

**Disclaimer**

This issue is part of the JOSS review in [https://github.com/openjournals/joss-reviews/issues/2969](https://github.com/openjournals/joss-reviews/issues/2969)
