### 2021-02-26 kuadrat/data-slicer#4 @sabinomaggi
> Furthermore, the newest version (0.0.12) gives you the functionality that you rightfully expected, namely loading a file directly when launching PIT from the command line.

Yes, now `pit` can load binary and text files from the comand line.
