### 2021-03-02 #2969 @kuadrat
Thanks @Chilipp and @sabinomaggi for your time, inputs and the statements of interest. It was indeed my observation that slicing data cubes to quickly inspect data is a task that is carried out in many fields. Ideally, users from different fields could write  plugins for their data formats and common processing routines to use `PIT` for just that task. In order for that to happen, many of your suggestions and raised issues are certainly of great importance (e.g. https://github.com/kuadrat/data-slicer/issues/10).

I'm excited to start working on these as soon as I can, which will not be before next week, however, due to an ongoing experiment.
