### 2021-03-08 093b60e
Added an alternative way to define keys for pickled dictionaries that is consistent with the Namespace approach and fixes #6.
