### 2021-03-08 7e80797
Documenting dependencies, fixing #8
