### 2021-03-08 7b7f66a
Added references to matplotlib, numpy and pyqtgraph, fixing #11
