### 2021-03-18 ef08c6f
Automated tests, as requested in #9
