+++ data_slicer/tests/pit.py
@@ -1,13 +1,0 @@
-"""
-Simply start up PIT with its example data.
-"""
-
-if __name__ == "__main__" :
-    from pyqtgraph.Qt import QtGui
-    from data_slicer.cmaps import ds_cmap
-    from data_slicer.pit import MainWindow
-
-    app = QtGui.QApplication([])
-    mw = MainWindow()
-    app.exec_()
-
