+++ data_slicer/tests/run_tests.py
@@ -1,0 +1,7 @@
+if __name__ == "__main__" :
+    import pytest
+    import pkg_resources
+
+    test_path = pkg_resources.resource_filename('data_slicer', 'tests/')
+    pytest.main([test_path])
+

[…]
