### 2021-03-18 089bb79
Updated requirements and Docs for automated tests, #9
