+++ requirements.txt
@@ -4,3 +4,5 @@ PyOpenGL>=3.1.1a1
 pyqtgraph>=0.11.0
 PyQt5>=5.13.2
 qtconsole>=4.6.0
+pytest>=6.2.2
+pytest-qt>=3.3.0
