### 2021-03-23 #2969 @kuadrat
> Hello @kuadrat, can you please let us know what the status of your updates is and how much more time you need?

Dear all,

of the 8 issues currently open, I have addressed 6 and referenced them in the respective commits, so the changes should be easy to find from within the issues.
I have not notified you for every individual change thinking that it would be easier for the reviewers to go through all of the issues at once.
I hope to be able to resolve the two issues left (https://github.com/kuadrat/data-slicer/issues/5 and https://github.com/kuadrat/data-slicer/issues/10) within the next two weeks.
