### 2021-03-23 #2969 @Chilipp
ah, thanks @kuadrat! I don't get notifications if you just reference the issue in the commit. Only if you also leave a comment in the issue, that's why I did not realize. I'll have a look into your updates asap
