### 2021-03-24 kuadrat/data-slicer#11 @Chilipp
alright, thanks @kuadrat. closing this as being fixed with 7b7f66a
