### 2021-03-24 kuadrat/data-slicer#7 @Chilipp
alright thanks @kuadrat! I consider this as fixed via 5efd421
