### 2021-03-24 kuadrat/data-slicer#8 @Chilipp
thanks @kuadrat. closing this as fixed through 7e80797
