### 2021-03-24 kuadrat/data-slicer#9 @Chilipp
hey @kuadrat! thanks for adding these tests to the software! Just a few small suggestions:

1. I strongly recommend that you add more in the near future and use a CI provider, such as Github Actions, CircleCI or Travis to automatically run your test suite on every commit.
2. I would not add `pytest` to the `requirements.txt` file as it is not necessary for using the software in production. Instead you should add it using the `extras_require` parameter in your `setup.py`, i.e. something like

   ```python
   setup(
       ...,
       extras_require={"tests": ["pytest>=6.2.2", "pytest-qt>=3.3.0"]},
   )
   ```
   or, even better, create a `tox.ini` file that lists the test commands and dependencies such that you only have to run `tox` to run all tests.

Anyway, these are just recommendations and no blocking issues for the review in https://github.com/openjournals/joss-reviews/issues/2969. I check the review items and leave this issue here open, but feel free to close it if you want.
