### 2021-03-24 #2969 @Chilipp
thanks @kuadrat for working on these issues! I updated my checklist @timtroendle. Let me know when you made progress with the remaining two issues @kuadrat.
