### 2021-03-24 #2969 @Chilipp
> @Chilipp Thank you for your review! Regarding your question: yes, there was an editorial scope check before. Still, we'd be interested in your opinion. Scholarly effort does not necessarily require to have been used or cited before, but there are other criteria as well.

alright @timtroendle, thanks for the clarification. It's still hard for me to judge this but concerning the potential use cases of the software as pointed out by me and @sabinomaggi (see _Whether the software is sufficiently useful that it is likely to be cited by other researchers working in this domain._), I'd say the scholarly effort for this software is high enough, so I check this item.
