### 2021-03-24 kuadrat/data-slicer#9 @kuadrat
It is me who has to thank you, @Chilipp, for the very useful suggestions (also with the other issues) and +1 for making me aware of ``tox``!
Now I definitely want to set things up with ``tox``. In my first tries with it today, however, it turned out to not be straightforward, likely due to Qt binding version conflicts, so I'll focus on the other issues for now ;)
