### 2021-03-26 ad7174a
Documentation on combining widgets & functions, issue #5
