+++ data_slicer/tests/test_threedwidget.py
@@ -9,6 +9,15 @@ import data_slicer.set_up_logging
 from data_slicer.widgets import ThreeDWidget
 
 def create_threedwindow() :
+    """ Create an example window that makes use of :class:`ThreeDWidget 
+    <data_slicer.widgets.ThreeDWidget>` to display a slice plane through a 3D 
+    data cube.
+
+    For convenience with 
+    :func:`~data_slicer.tests.test_threedwidget.test_threedwidget`, this 
+    function returns the created QWindow, the ThreeDWidget as well as the two 
+    QPushButtons.
+    """
     window = QtGui.QMainWindow()
     window.resize(800, 800)
 
@@ -22,24 +31,23 @@ def create_threedwindow() :
     w = ThreeDWidget()
     layout.addWidget(w, 0, 0, 1, 2)
 
+    # Create and add buttons
     button1 = QtGui.QPushButton()
     button1.setText('Reset')
     layout.addWidget(button1, 1, 0, 1, 1)
-
     button2 = QtGui.QPushButton()
     button2.setText('Randomize')
     layout.addWidget(button2, 1, 1, 1, 1)
 
     # Load example data
     data_path = pkg_resources.resource_filename('data_slicer', 'data/')
-#    datafile = 'testdata_100_150_200.p'
     datafile = 'pit.p'
     with open(data_path + datafile, 'rb') as f :
         data = pickle.load(f)
     shape = data.shape
     w.set_data(data)
 
-    # Button actions
+    # Define the button actions
     def reset_data() :
         w.set_data(data)
     button1.clicked.connect(reset_data)
