+++ doc/source/widgets.rst
@@ -1,0 +1,167 @@
+.. _sec-widgets:
+
+Widgets
+=======
+
+The widgets shipped with ``data-slicer`` can be re-combined to create new 
+applications.
+In order to do this, a certain level of familiarity with widget-based GUI 
+creation is required and some knowledge or experience with `Qt 
+<https://www.qt.io/>`_ or `PyQt <https://wiki.python.org/moin/PyQt>`_ is 
+recommended.
+Nevertheless, even without said experience, the `step-by-step 
+<sec-widget-tutorial>`_ tutorial below may help you figure out how to do such 
+things.
+
+List of available widgets
+-------------------------
+
+.. list-table:: 
+   :header-rows: 0
+   :widths: 40 60
+
+   * - :class:`data_slicer.imageplot.ImagePlot`
+     - A pseudocolor plot of a 2D dataset, similar to :class:`matplotlib 
+       pcolormesh <matplotlib.pyplot.pcolormesh>`.
+
+   * - :class:`data_slicer.imageplot.CrosshairImagePlot`
+     - An :class:`~data_slicer.imageplot.ImagePlot` with a draggable crosshair.
+
+   * - :class:`data_slicer.imageplot.CursorPlot`
+     - A regular data plot with a draggable line (cursor).
+
+   * - :class:`data_slicer.imageplot.Scalebar`
+     - A draggable scalebar. 
+
+   * - :class:`data_slicer.cutline.Cutline`
+     - A line that can be dragged on both ends and added to a
+       :class:`pyqtgraph.PlotWidget` to create arbitrary cuts.  
+
+   * - :class:`data_slicer.pit.MainWindow`
+     - The full main window of PIT, itself consisting of widgets from this 
+       list.  
+
+   * - :class:`data_slicer.widgets.ColorSliders`
+     - Gamma and vmax color sliders. Essentially just a pair of 
+       :class:`~data_slicer.imageplot.Scalebar` objects.
+
+   * - :class:`data_slicer.widgets.ThreeDWidget`
+     - A widget that allows showing xy-slices out of a 3D dataset as a plane in 
+       3D.
+
+   * - :class:`data_slicer.widgets.ThreeDSliceWidget`
+     - Subclass of :class:`~data_slicer.widgets.ThreeDWidget` that shows the 
+       xz- and yz-planes in addition to the xy plane.
+
+   * - :class:`data_slicer.widgets.FreeSliceWidget`
+     - Subclass of :class:`~data_slicer.widgets.ThreeDWidget` that makes use 
+       of a :class:`~data_slicer.cutline.Cutline` on an 
+       :class:`~data_slicer.imageplot.ImagePlot` to allow creating arbitrary 
+       slice-planes.
+
+.. _sec-widget-tutorial:
+
+Example
+-------
+
+In the following we will go through the creation of a simple App that uses a 
+:class:`~data_slicer.widgets.ThreeDSliceWidget` to display some data that can 
+be loaded when clicking a **Load** button.
+
+Step 1: Just a plain ThreeDSliceWidget
+......................................
+
+First, we take a look at what we would need to do in order to just create a 
+:class:`~data_slicer.widgets.ThreeDSliceWidget` without anything else.
+
+.. literalinclude:: ../examples/example_step1.py
+   :linenos:
+
+This should create a window with a 
+:class:`~data_slicer.widgets.ThreeDSliceWidget` - but there is no data 
+present inside the widget yet.
+
+Step 2: Adding a Load button
+............................
+
+In order to add a button, we need to change the structure of our application 
+a little bit.
+Since we need to be able to let Qt know where our different GUI elements 
[…]
