+++ doc/source/index.rst
@@ -28,6 +28,7 @@ and explain how it can be installed and used.
    plugins
    configuration
    ipython_crash_course
+   widgets
    examples
 
 .. toctree::
