### 2021-03-26 118322d
More explicit documentation on data access and setting, as requested in #4
