### 2021-03-26 24fd3bf
Extended information for contributors and developers, addressing #10
