+++ doc/source/contributing.rst
@@ -1,0 +1,45 @@
+.. _sec-contributing:
+
+Contributing
+============
+
+Thank you for helping to make this software package more useful!
+There are several ways in which you can contribute: giving feedback, 
+reporting bugs, issuing feature requests or fixing bugs and adding new 
+features yourself. 
+Furthermore, you can create and share your own :ref:`plugins <sec-plugin>`.
+
+Feedback, bugs and feature requests
+-----------------------------------
+
+The most straightforward and organized way to help improve this software is 
+by opening an **issue** on `the github 
+repository <https://github.com/kuadrat/data-slicer/issues>`.
+To do this, navigate to the **Issues** tab and click **New issue**.
+Please try to describe the bug you encountered or the new feature you would 
+like to see as detailed as possible.
+If you have several bugs/ideas please open a separate issue for each of them 
+in order to keep the discussions focused.
+
+If you have anything to tell me that does not seem to warrant opening an 
+issue or you simply prefer to contact me directly you can do this via e-mail:
+kevin.pasqual.kramer@protonmail.ch
+
+Contributing code
+-----------------
+
+If you have fixed a bug or created a new feature in the source code yourself, 
+it can be merged into this project.
+Code contributions will be acknowledged in this README or, if the number of 
+contributors grows too large, in a separate file.
+If you are familiar with the workflow on github, please go ahead and create a 
+pull request.
+If you are unsure you can always contact me via e-mail (see above).
+
+Plugins
+-------
+
+If you have created a PIT plugin, feel free to add it to the list on the 
+:ref:`plugins page <sec-plugin>`, either via pull request or through an 
+e-mail (see above).
+
