### 2021-03-26 #2969 @kuadrat
Dear @Chilipp and @sabinomaggi ,
I have now addressed all of the issues that have been opened during the review process so far and referenced them in the respective commits. In an attempt to reduce spam, I have not left additional comments in the issue trackers, hoping that the commit messages and ``diff``s will be self-explanatory. Do not hesitate to ask for explanations or specifications should anything be unclear when you check the different issues.
