### 2021-04-04 kuadrat/data-slicer#10 @Chilipp
hey @kuadrat! thanks for adding these community guidelines! A crucial thing concerning the scholarly effort of this package is the reusability and extensibility of your framework, and I think the contributing guide place an important role here. But there is another aspect that is still missing, that is a guide how to write a plugin. It's kind of a developer guide if you like. What are the essential steps? What classes should I inherit from if I want to write a plugin to the data-slicer package? Could you please include this information in your docs? From my perspective this can be held pretty short (e.g. something like https://psyplot.readthedocs.io/projects/psyplot-gui/en/latest/plugins.html), but it should be mentioned somewhere as an initial starting point.
