### 2021-04-06 #2969 @sabinomaggi
@timtroendle @kuadrat I have completed my reviewer's checklist above and, to me, the paper can now be accepted for publication. In particular I acknowledge the great work done by the authors in improving the documentation of the package.

I still think that the title of the paper should somehow mention `PIT` which, as I wrote in my report, is the "_real focal point of the paper_", but this should be taken as a suggestion since title issues are not included in the mandatory checklist.

`data-slicer` has truly the potential to become a very useful Python package, adressing the needs of a specific but very important community of scientists.
