### 2021-04-06 kuadrat/data-slicer#10 @kuadrat
Hi @Chilipp , this is exactly what I had in mind with issue #14 :) Since you explicitly mention it here, I will add the joss tag to it and start working on it next.
