### 2021-04-06 #2969 @Chilipp
hey @timtroendle, yes, I already commented in https://github.com/kuadrat/data-slicer/issues/10
