### 2021-04-06 387dae7
Doc: how-to make a plugin, fixing #14 and #10
