+++ doc/source/contributing.rst
@@ -14,7 +14,7 @@ Feedback, bugs and feature requests
 
 The most straightforward and organized way to help improve this software is 
 by opening an **issue** on `the github 
-repository <https://github.com/kuadrat/data-slicer/issues>`.
+repository <https://github.com/kuadrat/data-slicer/issues>`_.
 To do this, navigate to the **Issues** tab and click **New issue**.
 Please try to describe the bug you encountered or the new feature you would 
 like to see as detailed as possible.
@@ -30,7 +30,7 @@ Contributing code
 
 If you have fixed a bug or created a new feature in the source code yourself, 
 it can be merged into this project.
-Code contributions will be acknowledged in this README or, if the number of 
+Code contributions will be acknowledged in the README or, if the number of 
 contributors grows too large, in a separate file.
 If you are familiar with the workflow on github, please go ahead and create a 
 pull request.
@@ -42,4 +42,221 @@ Plugins
 If you have created a PIT plugin, feel free to add it to the list on the 
 :ref:`plugins page <sec-plugin>`, either via pull request or through an 
 e-mail (see above).
+
+.. _sec-plugin-how-to:
+
+Creating a plugin
+.................
+
+Creating a plugin is not difficult, but there are a few details that one 
+needs to get right in order for things to work.
+The package `ds-example-plugin <https://github.com/kuadrat/ds-example-plugin>`_
+was made to serve as an example to showcase how a plugin can be created.
+This section gives a step-by-step guide on how ``ds-example-plugin`` was 
+built and should allow you to figure out how to create your own plugins by 
+analogy.
+It is helpful to be familiar with the concept of object oriented 
+programming (working with classes and inheritance), but not mandatory.
+
+Before we start, let's have a look at what the example plugin does.
+You can either install it using ``pip install ds-example-plugin`` or download 
+the source code from github and save it (or a link to it) in the ``plugins`` 
+subdirectory of your :ref:`config directory <sec-config>`.
+Once downloaded and installed, start PIT and load the plugin::
+
+   [1] example = mw.load_plugin('ds_example_plugin')
+   Importing plugin ds_example_plugin (Example plugin).
+   
+Running ``example.help()`` then gives us a list of available functions. 
+There are just two (besides the ``help`` function).
+Let's try them both::
+
+   [2] example.example_function()
+   This message is sent to you by the Example plugin.
+
+   [3] example.blur()
+
+The first one doesn't do anything but print some text to the console.
+The second one blurs the data.
+It can also be called with a numeric argument to specify the *amount* of 
+blurring.
+Since the blurring is applied directly to the data, it can only be undone by 
+using ``pit.reset_data()``.
+
+That's really all this plugin does.
+Now let's see how it is implemented.
+In the following, we refer to the source code files that are found `on the 
+github <https://github.com/kuadrat/ds-example-plugin>`_.
+
+Step 1: Writing the Plugin class
+********************************
+
+The file ``ds_example_plugin.py`` contains the core of our plugin.
+Essentially, we define a ``class`` that inherits from 
+:class:`data_slicer.plugin.Plugin`::
+
+   from data_slicer import plugin
+
+   class Example_Plugin(plugin.Plugin) :
+      [...]
+
+The ``__init__()`` method of this class **must** call the superclass' 
+``__init__()`` through the line::
+
+   def __init__(self, *args, **kwargs) :
+      [...]
+      super().__init__(*args, **kwargs)
+      [...]
+
+It is recommended that you give your plugin a ``name`` and ``shortname`` by 
+setting the respective attributes in the ``__init__()`` function::
+
+   def __init__(self, *args, **kwargs) :
+      [...]
+      self.name = 'Example plugin'
+      self.shortname = 'example'
+
+Of course you can add any kind of code you need to in the ``__init__()`` 
+function.
[…]
