+++ doc/source/plugins.rst
@@ -4,7 +4,7 @@ Plugins
 =======
 
 It is possible for people to write plugins that add additional functionality 
-to PIT.
+to PIT (see also :ref:`the plugin writing guide <sec-plugin-how-to>`).
 A plugin for PIT is just a python module that connects to the backside of PIT.
 This module should either be found by the python version you are running 
 (i.e. you should be able to ``import`` it) or you can put it in your 
@@ -29,6 +29,8 @@ You can set it up such that your favourite plugins will always be loaded on
 startup of PIT.
 See :ref:`here <sec-config-plugins>`.
 
+.. _sec-plugin-list:
+
 List of known plugins
 ---------------------
 
@@ -42,6 +44,11 @@ either via pull request or through an e-mail
 
    * - Plugin Name (link) 
      - Description
+
+   * - `ds-example-plugin <https://github.com/kuadrat/ds_example_plugin>`_
+     - exists as a minimal example that can be used for guidance when 
+       creating your own plugins. A step-by-step tutorial on how it was made 
+       can be found :ref:`here <sec-plugin-how-to>`.
 
    * - `ds-arpes-plugin <https://github.com/kuadrat/ds_arpes_plugin>`_
      - tools for angle-resolved photoelectron spectroscopy (ARPES) data
