### 2021-04-06 kuadrat/data-slicer#10 @kuadrat
Hi again, @Chilipp, 387dae7 should address your latest suggestion. It turned out a little longer than expected, hopefully making things quite clear. I'd be happy to hear your thoughts!
