### 2021-04-12 kuadrat/data-slicer#10 @Chilipp
wow! that's indeed very detailed :smile: Thanks @kuadrat!
