### 2021-04-15 #2969 @kuadrat
@kyleniemeyer I have uploaded a zip of the current repository on Zenodo. The DOI is:

10.5281/zenodo.4692467
