### 2021-04-16 #2969 @kuadrat
@kyleniemeyer indeed, the invalid DOIs appear in the .bib file but are not actually used (and are not needed) in the paper.
Can I just remove them and add the ParaView DOI?

Also, I'll get back to you regarding the proof in about an hour at the latest.
