### 2021-04-16 #2969 @kyleniemeyer
@kuadrat no need to do anything about the invalid DOIs, it will not affect the final paper (just adds a bit of noise to that message). But yes, please do add the ParaView DOI
