+++ paper/paper.bib
@@ -59,7 +59,8 @@
   author = {Ahrens, J. and Geveci, Berk and Law, Charles},
   date = {2005-01-01},
   journaltitle = {Visualization Handbook},
-  title = {{{ParaView}}: {{An End}}-{{User Tool}} for {{Large Data Visualization}}}
+  title = {{{ParaView}}: {{An End}}-{{User Tool}} for {{Large Data Visualization}}},
+  doi = {10.1016/b978-012387582-2/50038-1}
 }
 
 @online{3dslicer,
