### 2021-04-16 #2969 @kuadrat
@kyleniemeyer Alright, thanks for clarifying. I have added the ParaView DOI and the preview did compile correctly. The proof looks fine to us as well.
