### 2021-04-19 #2969 @kuadrat
@kyleniemeyer Thank you for finishing up the publication process and thanks to @timtroendle for ediiting!

And many thanks again to @sabinomaggi and @Chilipp for their work, patience and useful inputs on improving the paper & code!
I must say that this has to be the most constructive review process I have witnessed so far.
